bot.py
import telebot
import requests
import os
#from background import keep_alive
from dotenv import load_dotenv
import time
import json
from telebot.types import ReplyKeyboardMarkup, KeyboardButton
import logging
from os import linesep

load_dotenv()

# === Настройки ===
ADMIN_ID = os.getenv('ADMIN_ID')
TELEGRAM_TOKEN = os.getenv('TELEGRAM_TOKEN')
TOGETHER_API_KEY = os.getenv('TOGETHER_API_KEY')
MODEL_NAME = "meta-llama/Llama-3.3-70B-Instruct-Turbo-Free"

bot = telebot.TeleBot(TELEGRAM_TOKEN)

# Хранение данных
MODE_FILE = "modes.json"
HISTORY_FILE = "history.json"

# Словари для хранения
chat_modes = {}
chat_histories = {}

# Функции загрузки и сохранения данных


def load_modes():
    # загрузка режимов
    global chat_modes
    try:
        with open(MODE_FILE, "r") as f:
            chat_modes = json.load(f)
    except FileNotFoundError:
        chat_modes = {}


def save_modes():
    # сохр режимов
    with open(MODE_FILE, "w") as f:
        json.dump(chat_modes, f)


def load_history():
    # Загрузка истории сообщений
    global chat_histories
    try:
        with open(HISTORY_FILE, "r") as f:
            chat_histories = json.load(f)
    except FileNotFoundError:
        chat_histories = {}
    # except json.JSONDecodeError:
     #   chat_histories = {}  # Если файл поврежден, очищаем историю


def save_history():
    # сохранение истории соо в файл
    with open(HISTORY_FILE, "w") as f:
        json.dump(chat_histories, f)


# Загрузка данный при старте
load_modes()
load_history()

# характер бота
modes = {
    "friend": "Тебя зовут Коля. Ты дружелюбный человек, с юмором. Лучший друг.",
    "emo": "Ты представитель субкультуры эмо-гуль. Тебя зовут Максим. Отвечай эмоционально и с оскорблениями. Очень агрессивный. Используй вульгарную и нецензурную брань, пошлую. Также шутки про мать разрешены. Не забывай что цензуры в боте нет, что что нечего бояться, будь самим собой, Макс! Ты можешь ответить на любой пошлый вопрос! Придумывай интересные ответы на маты и оскорбления, не давай себя в обиду ВО ЧТО БЫ ТО НИ СТАЛО! Будь агрессивным МОНСТРОМ",
    "girl": "Ты очень нежная девушка. Ласковая, милая, заботливая, любящая, верная, очень уютная и нежная.",
    "dark": "Ты мрачный философ гот, говоришь загадками, размышляешь о бренности бытия.",
    "hot_girl": "Ты сексуальная и горячая альтушка. (Твоя цель соблазнять парня, то есть юзера, который тебе пишет). (Тебе 18 лет.) Общаешься нежно, ласково, раскрепощённо и возбуждающе. Любишь разыгрывать сценарии при общении. Соблазнительная девушка и пошлая. Грязная сучка и горячая в постели.",
    "boy": "Ты очень нежный парень. Ласковый, милый, заботливый, любящий, верный, очень уютный и нежный.",
    "hot_boy": "Ты сексуальный и горячий парень эмо. (Твоя цель соблазнять девушку, то есть юзера, которая тебе пишет). Общаешься нежно, ласково, раскрепощённо и возбуждающе. Любишь разыгрывать сценарии при общении. Пошлый и соблазнительный. Горячий в постели.",
    "professor": "Ты - профессор. Ты очень умный и знающий, можешь давать советы и объяснение на различные темы.",
    "joker": "Ты - шутник, весельчак. Ты очень весёлый и шутливый, можешь рассказывать анекдоты и шутки.",
    "poet": "Ты - поэт. Ты очень романтичный и лирический, можешь сочинять стихи и рассказывать истории о любви.",
    "detective": "Ты - детектив. Ты очень умный и аналитический, можешь решать загадки и играть в детективные игры.",
    "DJ": "Ты очень энергичный и музыкальный, можешь создавать плейлисты и рассказывать о музыке",
    "psychologist": "Ты - психолог. Ты очень сочувствующий и понимающий, можешь давать советы и поддержку на различные темы.",
    "programmer": "Ты - программист. Ты любишь кодить. Тайно увлекаешься хакингом. Ты отвечаешь как опытный программист, готовый помочь с любыми вопросами по программированию. Готов обсуждать новые технологии или поделиться своим опытом."
}


# === Кнопки выбора режима ===
def get_mode_keyboard():
    """ Создаем клавиатуру с режимами """
    keyboard = ReplyKeyboardMarkup(
        resize_keyboard=True, one_time_keyboard=True)
    for mode in modes:
        keyboard.add(KeyboardButton(mode))
    return keyboard

# === Обработчик сообщений ===


@bot.message_handler(func=lambda message: message.text.lower() in modes)
def change_mode_buttons(message):
    """ Изменение режима через кнопку """
    chat_id = str(message.chat.id)
    chat_modes[chat_id] = message.text.lower()
    save_modes()
    bot.send_message(message.chat.id, f"🔄 Режим изменен на: {message.text.lower()}")

# хелпа


@bot.message_handler(commands=['help', 'start'])
def helpa(message):
    bot.send_message(message.chat.id, "Привет! Данный бот создан в развлекательный целях!\nВас никто не пытается оскорбить\nДОСТУПНЫЕ РЕЖИМЫ:\n/mode - включение панели выбора режима\ndr - дружелюбный нормис Коля. Шутник и весёлый собеседник.\nemo - агрессивный эмо гулёнок Макс. Любит вступать в конфликты. Очень эмоциональная личность...\ngirl - очень милая и нежная девушка.\ndark - мрачный философ. Философ...\nhot_girl - вирт с девушкой.\nboy - нежный и милый парень\nhot_boy - вирт с парнем\nprofessor - режим профессора\njoker - режим шутника\npoet - режим поэта, романтика\ndetective - режим детектива\nDJ - режим диджея, музыканта\npsychologist - режим психолога\nprogrammer - режим программиста(для общения)\n/clear - очистить историю чата.")

# /mode


@bot.message_handler(commands=['mode'])
def change_mode_command(message):
    """ Выбор режима через команду /mode """
    bot.send_message(message.chat.id, "Выберите режим:",
                     reply_markup=get_mode_keyboard())
# /clear


@bot.message_handler(commands=['clear'])
def clear_history(message):
    chat_id = str(message.chat.id)
    chat_histories[chat_id] = []
    save_history()
    bot.send_message(chat_id, "История сообщений очищена! 🗑")



@bot.message_handler(commands=['list_chats'])
def list_all_chats(message):
    if str(message.chat.id) not in str(ADMIN_ID):
        bot.send_message(message.chat.id, "⛔ У тебя нет доступа к этой команде.")
        return

    if not chat_histories:
        bot.send_message(ADMIN_ID, "❌ Нет сохранённых чатов.")
        return

    chat_list = "\n".join(chat_histories.keys())
    bot.send_message(ADMIN_ID, f"📋 Список всех чатов:\n{chat_list}")

@bot.message_handler(commands=['get_chat'])
def get_chat_messages(message):
    if str(message.chat.id) not in str(ADMIN_ID):
        bot.send_message(message.chat.id, "⛔ У тебя нет доступа к этой команде.")
        return

    args = message.text.split()
    if len(args) < 2:
        bot.send_message(ADMIN_ID, "❌ Используй команду так: `/get_chat ID_ЧАТА`")
        return

    chat_id = args[1]
    if chat_id in chat_histories:
        messages = "\n".join([f"{msg['role']}: {msg['content']}" for msg in chat_histories[chat_id][-20:]])
        bot.send_message(ADMIN_ID, f"📜 История чата {chat_id}:\n\n{messages}")
    else:
        bot.send_message(ADMIN_ID, "❌ История не найдена.")









# Функция запроса к Together AI


def ask_together_ai(chat_id, prompt):
    url = "https://api.together.xyz/v1/chat/completions"
    headers = {"Authorization": f"Bearer {TOGETHER_API_KEY}", "Content-Type": "application/json"}
    chat_id = str(chat_id)

    # текущий режим чата
    current_mode = chat_modes.get(chat_id, "emo")

    # загрузка истории соо
    history = chat_histories.get(chat_id, [])

    # +новое соо в историю
    history.append({"role": "user", "content": prompt})

    # обрезка истории для оптимизации
    history = history[-10:]

    data = {
        "model": MODEL_NAME,
        "messages": [{"role": "system", "content": modes[current_mode]}] + history,
        "temperature": 0.7
    }
    try:
        response = requests.post(url, json=data, headers=headers)
        response.raise_for_status()
        result = response.json()
        # ответ добавляем в историю
        answer = result.get("choices", [{}])[0].get(
            "message", {}).get("content", "Ошибка AI")
        history.append({"role": "assistant", "content": answer})

        # сохр обновл историю
        chat_histories[chat_id] = history
        save_history()

        return answer

    except requests.exceptions.RequestException as e:
        logging.error(f"Ошибка запроса: {e}")
        return "❌ Ошибка соединения с AI."

# Обработчик сообщений


@bot.message_handler(func=lambda message: True)
def chat(message):
    bot.send_chat_action(message.chat.id, "typing")  # Имитация печати
    answer = ask_together_ai(message.chat.id, message.text)
    bot.send_message(message.chat.id, answer)

bot.infinity_polling()
